requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
scikit-learn==1.6.1
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# MLOps Tools
mlflow==2.9.2
deepchecks==0.17.3

# Kafka
kafka-python==2.0.2

# Elasticsearch
elasticsearch==8.11.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Data Visualization (pour notebooks)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
